+++ features/support/helpers/lib_path.rb
@@ -1,0 +1,7 @@
+module LibPath
+
+  def lib_path
+    File.join(File.dirname(__FILE__), '../../../lib')
+  end
+
+end