+++ features/support/step_definitions.rb
@@ -1,9 +1,3 @@
-require_relative "../../lib/cute_print"
-
-Before do
-  CutePrint.configure { |c| c.reset }
-end
-
 Given(/^a file with:$/) do |contents|
   @example = Example.new(contents)
   @example.run