+++ lib/cute_print/core.rb
@@ -1,0 +1,3 @@
+# Require this file to get everything except the monkey-patches.
+
+require_relative "cute_print"