+++ lib/cute_print/core_ext.rb
@@ -1,2 +1,4 @@
+# Load all monkey-patches
+
 require_relative "core_ext/irb"
 require_relative "core_ext/object"