+++ lib/cute_print/printer.rb
@@ -36,15 +36,7 @@ module CutePrint
       @location_format = :filename
     end
 
-    # Inspect and write one or more objects.
-    #
-    # If called without a block, prints the inspected arguments, one
-    # on a line.
-    #
-    # If called with a block, prints the source code of the block and
-    # the inspected result of the block.
-    #
-    # @return nil
+    # @see CutePrint.q
     def q(*values, &block)
       formatter = Formatter.new(
         method: __method__,
@@ -57,15 +49,7 @@ module CutePrint
       nil
     end
 
-    # Inspect and write one or more objects, with source location.
-    #
-    # If called without a block, prints the inspected arguments, one
-    # on a line.
-    #
-    # If called with a block, prints the source code of the block and
-    # the inspected result of the block.
-    #
-    # @return nil
+    # @see CutePrint.ql
     def ql(*values, &block)
       formatter = Formatter.new(
         method: __method__,
@@ -78,15 +62,7 @@ module CutePrint
       nil
     end
 
-    # Pretty-print and write one or more objects.
-    #
-    # If called without a block, pretty-prints the pretty-printed
-    # arguments, one on a line.
-    #
-    # If called with a block, prints the source code of the block and
-    # pretty-prints the result of the block.
-    #
-    # @return nil
+    # @see CutePrint.qq
     def qq(*values, &block)
       formatter = Formatter.new(
         method: __method__,
@@ -98,15 +74,7 @@ module CutePrint
       nil
     end
 
-    # Pretty-print and write one or more objects, with source location.
-    #
-    # If called without a block, pretty-prints the arguments, one on a
-    # line.
-    #
-    # If called with a block, prints the source code of the block and
-    # pretty-prints the result of the block.
-    #
-    # @return nil
+    # @see CutePrint.qql
     def qql(*values, &block)
       formatter = Formatter.new(
         method: __method__,