+++ lib/cute_print/format.rb
@@ -1,0 +1,2 @@
+require_relative "format/inspect"
+require_relative "format/pretty_print"