+++ lib/cute_print/format/inspect.rb
@@ -1,0 +1,12 @@
+module CutePrint
+  module Format
+    # @api private
+    class Inspect
+
+      def format(width, value)
+        [value.inspect + "\n"]
+      end
+
+    end
+  end
+end