+++ lib/cute_print/format/pretty_print.rb
@@ -1,0 +1,17 @@
+require 'pp'
+require 'stringio'
+
+module CutePrint
+  module Format
+    # @api private
+    class PrettyPrint
+
+      def format(width, value)
+        out = StringIO.new
+        PP.pp(value, out, width)
+        out.string.lines
+      end
+
+    end
+  end
+end